+++ test/fixtures/lineCommentTS.js
@@ -21,7 +21,6 @@ const lineCommentTS = {
     }
   ],
   end: 23,
-  jsdoc: null,
   jsdocBlocks: [],
   loc: {
     end: {